Issue 5496 - 64bit: possible ABI issue in mars.h for D runtime.

+++ src/rt/mars.h
@@ -12,6 +12,7 @@
  *          http://www.boost.org/LICENSE_1_0.txt)
  */
 #include <stddef.h>
+#include <sys/types.h>
 
 #if __cplusplus
 extern "C" {
@@ -30,7 +31,7 @@ typedef struct Interface
 {
     struct ClassInfo *classinfo;
     struct Vtbl vtbl;
-    int offset;
+    ptrdiff_t offset;
 } Interface;
 
 typedef struct Object
